backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
cachetools
futu-api
questrade-api
